+++ visionforge-core/src/commonMain/kotlin/hep/dataforge/vision/SimpleVisionGroup.kt
@@ -1,17 +1,13 @@
 package hep.dataforge.vision
 
-import hep.dataforge.meta.Config
 import hep.dataforge.names.NameToken
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 
 @Serializable
-@SerialName("group")
+@SerialName("vision.group")
 public class SimpleVisionGroup : AbstractVisionGroup() {
-
-    override var styleSheet: StyleSheet? = null
-
     @SerialName("children")
     private val _children = HashMap<NameToken, Vision>()
     override val children: Map<NameToken, Vision> get() = _children