+++ visionforge-core/src/commonMain/kotlin/hep/dataforge/vision/StyleSheet.kt
@@ -5,7 +5,10 @@ package hep.dataforge.vision
 import hep.dataforge.meta.*
 import hep.dataforge.names.Name
 import hep.dataforge.names.asName
-import kotlinx.serialization.*
+import kotlinx.serialization.KSerializer
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.Transient
+import kotlinx.serialization.UseSerializers
 import kotlinx.serialization.builtins.MapSerializer
 import kotlinx.serialization.builtins.serializer
 import kotlinx.serialization.descriptors.SerialDescriptor
@@ -15,7 +18,7 @@ import kotlinx.serialization.encoding.Encoder
 /**
  * A container for styles
  */
-@Serializable
+@Serializable(StyleSheet.Companion::class)
 public class StyleSheet private constructor(private val styleMap: MutableMap<String, Meta>) {
     @Transient
     internal var owner: Vision? = null
@@ -30,8 +33,9 @@ public class StyleSheet private constructor(private val styleMap: MutableMap<Str
     private fun Vision.styleChanged(key: String, oldStyle: Meta?, newStyle: Meta?) {
         if (styles.contains(key)) {
             //TODO optimize set concatenation
-            val tokens: Collection<Name> = ((oldStyle?.items?.keys ?: emptySet()) + (newStyle?.items?.keys ?: emptySet()))
-                .map { it.asName() }
+            val tokens: Collection<Name> =
+                ((oldStyle?.items?.keys ?: emptySet()) + (newStyle?.items?.keys ?: emptySet()))
+                    .map { it.asName() }
             tokens.forEach { parent?.propertyChanged(it) }
         }
         if (this is VisionGroup) {
@@ -73,8 +77,17 @@ public class StyleSheet private constructor(private val styleMap: MutableMap<Str
         set(key, newStyle.seal())
     }
 
-    @ExperimentalSerializationApi
-    @Serializer(StyleSheet::class)
+    public fun update(key: String, meta: Meta) {
+        val existing = get(key)
+        set(key, existing?.edit { this.update(meta) } ?: meta)
+    }
+
+    public fun update(other: StyleSheet) {
+        other.items.forEach { (key, value) ->
+            update(key, value)
+        }
+    }
+
     public companion object : KSerializer<StyleSheet> {
         private val mapSerializer = MapSerializer(String.serializer(), MetaSerializer)
         override val descriptor: SerialDescriptor get() = mapSerializer.descriptor
@@ -90,11 +103,6 @@ public class StyleSheet private constructor(private val styleMap: MutableMap<Str
         }
     }
 }
-
-public fun StyleSheet.update(styleMeta: Meta){
-    TODO()
-}
-
 
 /**
  * Add style name to the list of styles to be resolved later. The style with given name does not necessary exist at the moment.