+++ visionforge-core/src/commonMain/kotlin/hep/dataforge/vision/Vision.kt
@@ -1,6 +1,7 @@
 package hep.dataforge.vision
 
 import hep.dataforge.meta.*
+import hep.dataforge.meta.descriptors.Described
 import hep.dataforge.names.Name
 import hep.dataforge.names.asName
 import hep.dataforge.names.toName
@@ -9,14 +10,15 @@ import hep.dataforge.values.asValue
 import hep.dataforge.vision.Vision.Companion.TYPE
 import hep.dataforge.vision.Vision.Companion.VISIBLE_KEY
 import kotlinx.serialization.PolymorphicSerializer
-import kotlinx.serialization.Serializable
 import kotlinx.serialization.Transient
+import kotlin.properties.ReadWriteProperty
+import kotlin.reflect.KProperty
 
 /**
  * A root type for display hierarchy
  */
 @Type(TYPE)
-public interface Vision : Configurable {
+public interface Vision : Configurable, Described {
 
     /**
      * The parent object of this one. If null, this one is a root.
@@ -37,12 +39,7 @@ public interface Vision : Configurable {
     /**
      * Get property (including styles). [inherit] toggles parent node property lookup
      */
-    public fun getProperty(name: Name, inherit: Boolean): MetaItem<*>?
-
-    /**
-     * Ger a property including inherited values
-     */
-    override fun getItem(name: Name): MetaItem<*>? = getProperty(name, true)
+    public fun getProperty(name: Name, inherit: Boolean = true): MetaItem<*>?
 
     /**
      * Trigger property invalidation event. If [name] is empty, notify that the whole object is changed
@@ -63,7 +60,7 @@ public interface Vision : Configurable {
      * List of names of styles applied to this object. Order matters. Not inherited.
      */
     public var styles: List<String>
-        get() =  properties[STYLE_KEY]?.stringList?: emptyList()
+        get() = properties[STYLE_KEY]?.stringList ?: emptyList()
         set(value) {
             config[STYLE_KEY] = value
         }
@@ -71,7 +68,7 @@ public interface Vision : Configurable {
     /**
      * Update this vision using external meta. Children are not updated.
      */
-    public fun update(meta: Meta)
+    public fun update(change: Vision)
 
     public companion object {
         public const val TYPE: String = "vision"
@@ -90,15 +87,60 @@ public fun Vision.getProperty(key: String, inherit: Boolean = true): MetaItem<*>
     getProperty(key.toName(), inherit)
 
 /**
+ * A convenience method to pair [getProperty]
+ */
+public fun Vision.setProperty(key: Name, value: Any?) {
+    config[key] = value
+}
+
+/**
+ * A convenience method to pair [getProperty]
+ */
+public fun Vision.setProperty(key: String, value: Any?) {
+    config[key] = value
+}
+
+/**
  * Find a style with given name for given [Vision]. The style is not necessary applied to this [Vision].
  */
 public tailrec fun Vision.resolveStyle(name: String): Meta? =
     (this as? VisionGroup)?.styleSheet?.get(name) ?: parent?.resolveStyle(name)
 
-
 /**
  * Control visibility of the element
  */
 public var Vision.visible: Boolean?
-    get() = getItem(VISIBLE_KEY).boolean
-    set(value) = setItem(VISIBLE_KEY, value?.asValue())
+    get() = getProperty(VISIBLE_KEY).boolean
+    set(value) = config.setValue(VISIBLE_KEY, value?.asValue())
+
+/**
+ * Convinience delegate for properties
+ */
+public fun Vision.property(
+    default: MetaItem<*>? = null,
+    key: Name? = null,
+    inherit: Boolean = true,
+): MutableItemDelegate =
+    object : ReadWriteProperty<Any?, MetaItem<*>?> {
+        override fun getValue(thisRef: Any?, property: KProperty<*>): MetaItem<*>? {
+            val name = key ?: property.name.toName()
+            return getProperty(name, inherit) ?: default
+        }
+
+        override fun setValue(thisRef: Any?, property: KProperty<*>, value: MetaItem<*>?) {
+            val name = key ?: property.name.toName()
+            setProperty(name, value)
+        }
+    }
+
+//TODO replace by value
+fun Vision.properties(inherit: Boolean = true): MutableItemProvider = object : MutableItemProvider {
+    override fun getItem(name: Name): MetaItem<*>? {
+        return getProperty(name, inherit)
+    }
+
+    override fun setItem(name: Name, item: MetaItem<*>?) {
+        setProperty(name, item)
+    }
+
+}