+++ visionforge-solid/src/commonMain/kotlin/hep/dataforge/vision/solid/BasicSolid.kt
@@ -1,15 +1,20 @@
 package hep.dataforge.vision.solid
 
-import hep.dataforge.meta.*
-import hep.dataforge.vision.AbstractVision
+import hep.dataforge.meta.Meta
+import hep.dataforge.meta.descriptors.NodeDescriptor
+import hep.dataforge.meta.float
+import hep.dataforge.meta.get
+import hep.dataforge.meta.node
 import hep.dataforge.vision.Vision
+import hep.dataforge.vision.VisionBase
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.UseSerializers
 
 @Serializable
 @SerialName("solid")
-public open class BasicSolid: AbstractVision(), Solid {
+public open class BasicSolid : VisionBase(), Solid {
+    override val descriptor: NodeDescriptor get() = Solid.descriptor
+
     @Serializable(Point3DSerializer::class)
     override var position: Point3D? = null
 
@@ -19,16 +24,16 @@ public open class BasicSolid: AbstractVision(), Solid {
     @Serializable(Point3DSerializer::class)
     override var scale: Point3D? = null
 
-    override fun update(meta: Meta) {
+    override fun update(change: Vision) {
         fun Meta.toVector(default: Float = 0f) = Point3D(
             this[Solid.X_KEY].float ?: default,
             this[Solid.Y_KEY].float ?: default,
             this[Solid.Z_KEY].float ?: default
         )
 
-        meta[Solid.POSITION_KEY].node?.toVector()?.let { position = it }
-        meta[Solid.ROTATION].node?.toVector()?.let { rotation = it }
-        meta[Solid.SCALE_KEY].node?.toVector(1f)?.let { scale = it }
-        super.update(meta)
+        change.properties[Solid.POSITION_KEY].node?.toVector()?.let { position = it }
+        change.properties[Solid.ROTATION].node?.toVector()?.let { rotation = it }
+        change.properties[Solid.SCALE_KEY].node?.toVector(1f)?.let { scale = it }
+        super.update(change)
     }
 }