+++ visionforge-solid/src/commonMain/kotlin/hep/dataforge/vision/solid/PolyLine.kt
@@ -7,6 +7,7 @@ import hep.dataforge.names.Name
 import hep.dataforge.names.asName
 import hep.dataforge.names.plus
 import hep.dataforge.vision.VisionContainerBuilder
+import hep.dataforge.vision.properties
 import hep.dataforge.vision.set
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -17,7 +18,7 @@ import kotlinx.serialization.UseSerializers
 public class PolyLine(public var points: List<Point3D>) : BasicSolid(), Solid {
 
     //var lineType by string()
-    public var thickness: Number by number(1.0, key = SolidMaterial.MATERIAL_KEY + THICKNESS_KEY)
+    public var thickness: Number by properties().number(1.0, key = SolidMaterial.MATERIAL_KEY + THICKNESS_KEY)
 
     public companion object {
         public val THICKNESS_KEY: Name = "thickness".asName()