+++ visionforge-solid/src/commonMain/kotlin/hep/dataforge/vision/solid/Solid.kt
@@ -14,7 +14,6 @@ import hep.dataforge.vision.enum
 import hep.dataforge.vision.solid.Solid.Companion.DETAIL_KEY
 import hep.dataforge.vision.solid.Solid.Companion.IGNORE_KEY
 import hep.dataforge.vision.solid.Solid.Companion.LAYER_KEY
-import kotlinx.serialization.UseSerializers
 
 /**
  * Interface for 3-dimensional [Vision]
@@ -24,7 +23,7 @@ public interface Solid : Vision {
     public var rotation: Point3D?
     public var scale: Point3D?
 
-    override val descriptor: NodeDescriptor? get() = Companion.descriptor
+    override val descriptor: NodeDescriptor get() = Companion.descriptor
 
     public companion object {
         //        val SELECTED_KEY = "selected".asName()
@@ -123,7 +122,7 @@ public enum class RotationOrder {
  */
 public var Solid.rotationOrder: RotationOrder
     get() = getItem(Solid.ROTATION_ORDER_KEY).enum<RotationOrder>() ?: RotationOrder.XYZ
-    set(value) = setItem(Solid.ROTATION_ORDER_KEY, value.name.asValue())
+    set(value) = setValue(Solid.ROTATION_ORDER_KEY, value.name.asValue())
 
 
 /**
@@ -131,7 +130,7 @@ public var Solid.rotationOrder: RotationOrder
  */
 public var Solid.detail: Int?
     get() = getProperty(DETAIL_KEY, false).int
-    set(value) = setItem(DETAIL_KEY, value?.asValue())
+    set(value) = setValue(DETAIL_KEY, value?.asValue())
 
 /**
  * If this property is true, the object will be ignored on render.
@@ -139,7 +138,7 @@ public var Solid.detail: Int?
  */
 public var Vision.ignore: Boolean?
     get() = getProperty(IGNORE_KEY, false).boolean
-    set(value) = setItem(IGNORE_KEY, value?.asValue())
+    set(value) = setValue(IGNORE_KEY, value?.asValue())
 
 //var VisualObject.selected: Boolean?
 //    get() = getProperty(SELECTED_KEY).boolean