+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## [Unreleased]
 ### Added
 - Server module
+- Change collector
 
 ### Changed
+- Vision does not implement ItemProvider anymore. Property changes are done via `getProperty`/`setProperty` and `property` delegate.
 
 ### Deprecated